partcad/requirements.txt
build123d>=0.3.0
cadquery>=2.4.0
numpy>=1.25.2
scipy>=1.11.1
pyyaml>=6.0.1
GitPython>=3.1.40
ruamel.yaml>=0.18.5
packaging>=23.1
setuptools
jinja2
requests
google.generativeai
openai